Tweaked the linux installation. Moved yombo/core/tests to proper place.
Installer for Linux no longer creates a new user. This created two copies
and made it difficult to develope/test/run.  Now, the installer just points
to the current installation. This really only affects users who do tweak
with the gateway - makes it easier on them. Everyone else, it still runs
fine. The installer configures the daemon to run as the user that ran
the install script.

Move random tests in yombo/core/tests to tests folder, it's new home. Also
make a couple of tests for the auth module. Doesn't do much.