Created instalation directory and steamlining installation.
Installation was a chore, but it's starting be easy. Created/updated
scripts to handle most this. Also modified the yombo.sh script to be
able to run in daemon called from init, or called locally by the user.

This is only tested on Debian as that is all I have.