pep 8 day... turing into refactoring day.
SQLDict made into a full library. This should provide better functionality
down the road.

Many bug fixes found during pep 8 day.

Changed unloading to now wait for yields to complete. Before, it just
quit. This was found during the SQLDict revamp.

TODO: States API will remain the same, but the backend will need to be
overhauled. Possible along with atoms. The API's will remain the same,
but will be a direct database backend instead of using SQLDict...maybe.

Messages: This antique was around since before hooks were around
and was the ONLY method to send commands, status updates, events. Now
with hooks around, messages are not as important. Will either be
changes to looks like AMQP messages or will be dropped. Still deciding.

Should really be using HOOKS as it's vastly more efficient and effective
anyways. More hooks to come regardless if messages are here to stay or not.