Update:  Summary - Now handle refresh and session tokens for Yombo API access.
When a user logs in, a refresh token (long term) and a session token (short term) are created. They are stored
here, encrypted with AES. The key is the session ID and is only available during the request.

TODO: Check if session tokens need to be refreshed. See #4
TODO: Stop storing session long IDs, will only be available during the request. See #5.