Fix: the layout.tpl no longer loads sw.js. Also, check if session is defined.
The sw.js is apart of the nuxt framework and was causing errors on gateway setup.
This files was previously loaded to speed up loading the frontend later, but the
file is small, so it's not a big deal.

The layout.tpl also displays errors for the session, but would cause an
error if there was no session on anon-pages. A simple check is in place
to check if there is a session to pull alerts from.