Fix: A web session wasn't properly loaded.
This fixes the session loading properly. This also closes  #5 so that
long session id's are not stored in the session. The session long
id is what the browser sends as a key, but we use a shorter
ID for tracking and so that the long id cannot be easily obtained
by getting a copy of the database.

Also, the session long id is used to encrypt the users sensitive data
and is only available during a request.