Fix: Add route for sw.js and fix autologin to sw.js
sw.js should have a max age of 0 for the cache, so a new route
handler was created for this in user.py

In some wierd instances, the auto redirect on login would goto a url
such as a css, js, etc page. This isn't what the user intended. So,
now once a login redirect is set, it shouldn't be overridden. However,
if a login redirect to something silly like a css or js file, just
redirect to the homepage.