New: lang.py For bulk update language translations
When updating language translation for the Frontend application, it can
be time consuming to do it manually. This script wil look for
'Yombo_Frontend-locale.zip' as downloaded from lokalise.co.  This will
place them in the frontend/lang folder all processed and ready to
go.  Don't forget to eat your ice cream.