B.A.C. - Standardizing data sync, database access, devices have states.
This is a bunch of little commits rolled into one BAC to make life easier.

== Data sync ==
There's been a long running issue of syncing data between runtime data, the
database, and Yombo API. The biggest portion of this commit is that now if runtime
data is changed, data will automatically be synced between all three.

=== Sources of data ===
The primary 'truth' of data is the Yombo API cloud. From here, the gateway downloads
it's run time data - such as devices, commands, what modules should be installed,
etc. When the gateway is started, this data is first downloaded directly to the
sqlite database. After this is complete, the gateway will continue loading libraries,
while each library then loads it's data from the database.

=== Source of data change ===
Data can be changed during runtime, which should cause the gateway to become out
of sync. Now data can be changed during runtime within the gateway and that data
will be sync'd to the database and if needed, uploaded to Yombo API.

Addtionally, data can be changed at Yombo API and needs to be sent to the Gateway.
This change will not make this possible in a future commit. In the future, any
changes will be sent VIA amqp live and the gateway will be able to react.

=== To do ===
This commit doesn't complete the actual last step - sending the data to Yombo API
cloud. This commit creates the framework for another possible semi-BAC to be done
to consolidate all Yombo API requests into the Yombo API library.

== Database Access ==
In order to support data sync, work has been started to standardize the database
library directory. This is not complete, but it's working just enough for this
BAC to be done. Further work is needed to standardize the remaining tables.

== Device States ==
Devices now have states instead of status. A device is in a state, that is, it can
be on, off, running, playing, opening, opened, etc. Everything has been renamed
across all Yombo support modules as well.

== Good by silly caches ==
Various caches were implemented to speed up accessing variables for modules and
devices. The data sync has allowed this to removed. All data is now loaded into
memory and synced to the database as needed. Caches are no longer needed nor
do they need to be maintained.

This doesn't lessen the memory requirements, it should be about even. The caches
would contain redundant data, which is no longer the case. However, the data
is now represented as instances.

== To do ==
After this BAC, the following todo's were created:
* Complete the Yombo API uploading of data.
* Server side - send applicable changes from Yombo cloud to gateways.
* Gateway AMQP client needs to be ehanced/extended to send these changes
  to the various locations where data is to be changed.