B.A.C.: Standardize mixins, use the Entity class, standards everywhere.
The mixin's were a complete mess on ordering and was was sent in the with the __init__.
This has now be mostly standardized. This allows adding more mixins with ease.

The Entity class has consolidated YomboBaseMixin and MagicAttributes. Global magic
attributes and functions should now be added here.  The Entity class should be the
parent of every class within the system, when possible. A future commit will allow
all Entity classes to be listed in a single place - great for debugging and tweaking.

Fixed the Frontend build process. The builder bash script has a little more smarts
and will copy the frontend for production builds automatically to the final
destination. One less thing to manage and track within the framework.

Calling hooks globaly has been removed from the utils helper into it's own
file: yombo.utisl.hookinvoke

Started work on a global YomboAPI framework for making changes back to Yombo
API. This will be used as part of the sync to everywhere mixin.